refactored options management
