+++ tests/fr/inria/verveine/extractor/java/VerveineJTest_AdHoc.java
@@ -51,15 +51,15 @@ public class VerveineJTest_AdHoc extends VerveineJTest_Basic {
 	 */
 	@Before
 	public void setUp() throws Exception {
-		new File(VerveineJParser.OUTPUT_FILE).delete();
+		new File(VerveineJOptions.OUTPUT_FILE).delete();
 		parser = new VerveineJParser();
 		repo = parser.getFamixRepo();
 	}
 
 	private void parse(String[] sources) {
-		parser.setOptions(sources);
+		parser.configure( sources);
 		parser.parse();
-		parser.emitMSE(VerveineJParser.OUTPUT_FILE);
+		parser.emitMSE(VerveineJOptions.OUTPUT_FILE);
 	}
 
 	@Test
