+++ tests/fr/inria/verveine/extractor/java/VerveineJTest_CommentsMethod.java
@@ -1,19 +1,18 @@
 package fr.inria.verveine.extractor.java;
 
 
-import eu.synectique.verveine.core.gen.famix.FileAnchor;
-import eu.synectique.verveine.core.gen.famix.IndexedFileAnchor;
-import eu.synectique.verveine.core.gen.famix.Method;
-import org.junit.Before;
-import org.junit.Test;
+import static org.junit.Assert.assertEquals;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
 
-import static org.junit.Assert.*;
+import org.junit.Before;
+import org.junit.Test;
+
+import eu.synectique.verveine.core.gen.famix.IndexedFileAnchor;
+import eu.synectique.verveine.core.gen.famix.Method;
 
 public class VerveineJTest_CommentsMethod extends VerveineJTest_Basic {
 
@@ -26,16 +25,16 @@ public class VerveineJTest_CommentsMethod extends VerveineJTest_Basic {
      */
     @Before
     public void setUp() throws Exception {
-        new File(VerveineJParser.OUTPUT_FILE).delete();
+        new File(VerveineJOptions.OUTPUT_FILE).delete();
         parser = new VerveineJParser();
         repo = parser.getFamixRepo();
     }
 
-    private void parse(String[] sources) {
-        parser.setOptions(sources);
-        parser.parse();
-        parser.emitMSE(VerveineJParser.OUTPUT_FILE);
-    }
+	private void parse(String[] sources) {
+		parser.configure( sources);
+		parser.parse();
+		parser.emitMSE(VerveineJOptions.OUTPUT_FILE);
+	}
 
     @Test
     public void testStartPosSourceAnchorMethod() {
