+++ tests/fr/inria/verveine/extractor/java/VerveineJTest_Exceptions.java
@@ -27,10 +27,10 @@ public class VerveineJTest_Exceptions extends VerveineJTest_Basic {
      */
     @Before
     public void setUp() throws Exception {
-        new File(VerveineJParser.OUTPUT_FILE).delete();
+        new File(VerveineJOptions.OUTPUT_FILE).delete();
         VerveineJParser parser = new VerveineJParser();
         repo = parser.getFamixRepo();
-        parser.setOptions(new String[] {"test_src/exceptions"});
+        parser.configure( new String[] {"test_src/exceptions"});
         parser.parse();
         //parser.emitMSE(VerveineJParser.OUTPUT_FILE);
     }
