+++ tests/fr/inria/verveine/extractor/java/VerveineJTest_Generics.java
@@ -22,10 +22,10 @@ public class VerveineJTest_Generics extends VerveineJTest_Basic {
      */
     @Before
     public void setUp() throws Exception {
-        new File(VerveineJParser.OUTPUT_FILE).delete();
+        new File(VerveineJOptions.OUTPUT_FILE).delete();
         VerveineJParser parser = new VerveineJParser();
         repo = parser.getFamixRepo();
-        parser.setOptions(new String[] {"test_src/generics"});
+        parser.configure( new String[] {"test_src/generics"});
         parser.parse();
         //parser.emitMSE(VerveineJParser.OUTPUT_FILE);
     }
