+++ tests/fr/inria/verveine/extractor/java/VerveineJTest_Inner.java
@@ -27,16 +27,16 @@ public class VerveineJTest_Inner extends VerveineJTest_Basic {
      */
     @Before
     public void setUp() throws Exception {
-        new File(VerveineJParser.OUTPUT_FILE).delete();
+        new File(VerveineJOptions.OUTPUT_FILE).delete();
         parser = new VerveineJParser();
         repo = parser.getFamixRepo();
     }
 
-    private void parse(String[] sources) {
-        parser.setOptions(sources);
-        parser.parse();
-        parser.emitMSE(VerveineJParser.OUTPUT_FILE);
-    }
+	private void parse(String[] sources) {
+		parser.configure( sources);
+		parser.parse();
+		parser.emitMSE(VerveineJOptions.OUTPUT_FILE);
+	}
 
     @Test
     public void testNumberOfClass() {
