+++ tests/fr/inria/verveine/extractor/java/VerveineJTest_JWSBasic.java
@@ -38,10 +38,10 @@ public class VerveineJTest_JWSBasic extends VerveineJTest_Basic {
 	 */
 	@Before
 	public void setUp() throws Exception {
-		new File(VerveineJParser.OUTPUT_FILE).delete();
+		new File(VerveineJOptions.OUTPUT_FILE).delete();
 		VerveineJParser parser = new VerveineJParser();
 		repo = parser.getFamixRepo();
-		parser.setOptions(new String[] {"test_src/jws_basic"});
+		parser.configure( new String[] {"test_src/jws_basic"});
 		parser.parse();
 	//	parser.emitMSE(VerveineJParser.OUTPUT_FILE);
 	}
