+++ tests/fr/inria/verveine/extractor/java/VerveineJTest_Summarized.java
@@ -88,7 +88,7 @@ public class VerveineJTest_Summarized extends VerveineJTest_Basic {
 	 */
 	@Before
 	public void setUp() throws Exception {
-		new File(VerveineJParser.OUTPUT_FILE).delete();
+		new File(VerveineJOptions.OUTPUT_FILE).delete();
 
 		String[] files = new String[] {
 				"AbstractDestinationAddress.java",
@@ -126,11 +126,11 @@ public class VerveineJTest_Summarized extends VerveineJTest_Basic {
 		
 		VerveineJParser parser = new VerveineJParser();
 		repo = parser.getFamixRepo();
-		parser.setOptions(args);
+		parser.configure( args);
 		parser.parse();
 		
-		new File(VerveineJParser.OUTPUT_FILE).delete();  // delete old MSE file
-		parser.emitMSE(VerveineJParser.OUTPUT_FILE);  // to create a new one
+		new File(VerveineJOptions.OUTPUT_FILE).delete();  // delete old MSE file
+		parser.emitMSE(VerveineJOptions.OUTPUT_FILE);  // to create a new one
 	}
 
 	@Test(timeout=100)
